src/keeper-portable/requirements.txt
slack-bolt>=1.18.0
anthropic>=0.25.0
httpx>=0.27.0
python-dotenv>=1.0.0
uvicorn>=0.29.0
